Changed discusses things before pull request
Re-worked available standard SPIL events (removed/added some).

Also added Sebastian's packages and promotions code and tested it with
the JSONHelper code to deserialize/serialize it between JSON<>Object.
Both packages and promotions seem to work just fine!

I did not add exception handling for user-made JSON and classes yet but
did add a comment that we may add this in the future. This shouldn't be
a problem, developers can use the standard exceptions to debug their
JSON and classes.